requirements.txt
streamlit
llama-index
llama-index-llms-gemini
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
python-dotenv
